+++ R/DSF_Downsample.R
@@ -72,7 +72,6 @@ DSF_Downsample <- function(dsd, factor = 1) {
 #' @export
 get_points.DSF_Downsample <- function(x,
   n = 1L,
-  outofpoints = c("stop", "warn", "ignore"),
   info = TRUE,
   ...) {
   .nodots(...)
@@ -86,7 +85,6 @@ get_points.DSF_Downsample <- function(x,
   d <-
     get_points(x$dsd,
       n = n_take,
-      outofpoints = outofpoints,
       info = info,
       ...)